refactoring: #845 move module dict_helpers to utils package
